lrp: move api handler from daemon to lrp hive cell
Currently, the daemon implements the LRP API handler. With it
comes the last dependency to the LRP manager from the daemon.

Therefore, this commit moves the LRP API handler from the daemon
to the existing LRP hive cell and removes the dependency from the
daemon to the LRP manager.
