+++ pkg/redirectpolicy/api.go
@@ -1,23 +1,26 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package cmd
+package redirectpolicy
 
 import (
 	"github.com/go-openapi/runtime/middleware"
 
 	"github.com/cilium/cilium/api/v1/models"
-	. "github.com/cilium/cilium/api/v1/server/restapi/service"
+	"github.com/cilium/cilium/api/v1/server/restapi/service"
 	"github.com/cilium/cilium/pkg/logging/logfields"
-	"github.com/cilium/cilium/pkg/redirectpolicy"
 )
 
-func getLRPHandler(d *Daemon, params GetLrpParams) middleware.Responder {
-	log.WithField(logfields.Params, logfields.Repr(params)).Debug("GET /lrp request")
-	return NewGetLrpOK().WithPayload(getLRPs(d.lrpManager))
+type getLrpHandler struct {
+	lrpManager *Manager
 }
 
-func getLRPs(rpm *redirectpolicy.Manager) []*models.LRPSpec {
+func (h *getLrpHandler) Handle(params service.GetLrpParams) middleware.Responder {
+	log.WithField(logfields.Params, logfields.Repr(params)).Debug("GET /lrp request")
+	return service.NewGetLrpOK().WithPayload(getLRPs(h.lrpManager))
+}
+
+func getLRPs(rpm *Manager) []*models.LRPSpec {
 	lrps := rpm.GetLRPs()
 	list := make([]*models.LRPSpec, 0, len(lrps))
 	for _, v := range lrps {
