+++ pkg/redirectpolicy/cell.go
@@ -6,6 +6,7 @@ package redirectpolicy
 import (
 	"github.com/cilium/hive/cell"
 
+	serviceapi "github.com/cilium/cilium/api/v1/server/restapi/service"
 	agentK8s "github.com/cilium/cilium/daemon/k8s"
 	"github.com/cilium/cilium/pkg/endpointmanager"
 	"github.com/cilium/cilium/pkg/k8s"
@@ -18,6 +19,7 @@ var Cell = cell.Module(
 	"LRP Manager",
 
 	cell.Provide(newLRPManager),
+	cell.Provide(newLRPApiHandler),
 )
 
 type lrpManagerParams struct {
@@ -31,4 +33,10 @@ type lrpManagerParams struct {
 
 func newLRPManager(params lrpManagerParams) *Manager {
 	return NewRedirectPolicyManager(params.Svc, params.SvcCache, params.Lpr, params.Ep)
+}
+
+func newLRPApiHandler(lrpManager *Manager) serviceapi.GetLrpHandler {
+	return &getLrpHandler{
+		lrpManager: lrpManager,
+	}
 }
